+++ CHANGELOG.rst
@@ -3,6 +3,9 @@
 
 ### v1.1.1 (unreleased)
 
+- Implement cache to prevent scanning multiple times a duplicated codebase resource.
+  https://github.com/nexB/scancode.io/issues/70
+
 - Create the virtualenv using the virtualenv.pyz app in place of the bundled "venv".
   https://github.com/nexB/scancode.io/issues/104